Add security config

+++ src/main/java/datamartapp/repositories/RoleRepository.java
@@ -1,9 +1,9 @@
 package datamartapp.repositories;
 
-import datamartapp.model.Role;
+import datamartapp.model.users.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 }
